Restructure et_replay sub-project
et_replay/
  - several high level modules mainly responsible for parsing
  - tools # previously was on a layer above, moved here to simplify
    imports management
  - comm # comm utils and functionalities, to be expanded
  - __init__.py # to hide files locations yet always export important
    symbols

`train` now doesn't care about actual files' locations.
