+++ et_replay/lib/__init__.py
@@ -1,0 +1,3 @@
+from .execution_trace import ExecutionTrace
+
+__all__ = ["ExecutionTrace"]
