+++ et_replay/lib/comm/pytorch_backend_utils.py
@@ -8,10 +8,9 @@ from abc import ABC, abstractmethod
 from typing import Dict, List, Optional
 
 import torch
-
-from et_replay.lib.comm.param_profile import paramTimer
-
 from torch.distributed import ProcessGroup
+
+from .param_profile import paramTimer
 
 logger = logging.getLogger(__name__)
