+++ et_replay/pyproject.toml
@@ -7,8 +7,8 @@ name = "et_replay"
 version = "0.5.0"
 
 [tool.setuptools.package-dir]
-"et_replay.lib.comm" = "lib/comm"
-"et_replay.tools" = "tools"
+"et_replay" = "lib"
+"et_replay.tools" = "lib/tools"
 
 [project.scripts]
 comm_replay = "et_replay.tools.comm_replay:main"
